test(llm): split Responses adapter coverage
